+++ GuryevGrigory/Laba2/Laba.c/Functions.h
@@ -2,12 +2,12 @@
 #include <stdio.h>
 #include <math.h>
 
-double next_element(char func_type, double x, double prev_element, int i);
+float next_element(char func_type, float x, float prev_element, int i);
 
-double sum(char func_type, double x, double first_element);
+float sum(char func_type, float x, float first_element);
 
-static double recursion_sum(char func_type, double prev_elem, double x, int i);
+static float recursion_sum(char func_type, float prev_elem, float x, int i);
 
-double invers_sum(char func_type, double first_elem, double x);
+float invers_sum(char func_type, float first_elem, float x);
 
-double pairwise_sum(char func_type, double first_elem, double x);
+float pairwise_sum(char func_type, float first_elem, float x);
